test: move the live LiteLLM client tests where they can reach a provider
The 24 tests in test_litellm_client.py that call OpenAI and Anthropic sat
under tests/server/llm/, so litellm.completion was patched under them:
_is_e2e_test_run exempts only paths containing e2e_tests, and nothing
about a marker or an API key changes that. They asserted against canned
mock text no matter how they were invoked.

Split the file along the line that already existed. The offline half
stays -- config resolution, API-key routing across providers, Azure
endpoint fields, image encoding, all of which need no credential. The
live half moves to tests/e2e_tests/test_litellm_client_real_llm.py and
picks up an autouse guard asserting MOCK_LLM_RESPONSE is unset, so a
future move back reports itself instead of silently passing.

Verified the bypass: with a syntactically valid but bogus OPENAI_API_KEY,
the moved test now reaches OpenAI and fails with AuthenticationError
("Incorrect API key provided"). From the old path the same run returned
mock text. Real assertions against real providers are unverified here --
this machine has no OpenAI or Anthropic key.

Shared helpers (create_minimal_png, MathResult) are imported from the
offline module, matching how test_litellm_strict_structured_compat_e2e
imports from its non-e2e sibling.
